refactor: 아티스트 엔티티 공통 폴더로 관리
