+++ src/main/java/com/prography/lighton/artist/admin/application/ManageArtistApplicationService.java
@@ -1,8 +1,8 @@
 package com.prography.lighton.artist.admin.application;
 
 import com.prography.lighton.artist.admin.presentation.ManageArtistApplicationRequestDTO;
-import com.prography.lighton.artist.users.domain.entity.Artist;
-import com.prography.lighton.artist.users.domain.entity.enums.ApproveStatus;
+import com.prography.lighton.artist.common.domain.entity.Artist;
+import com.prography.lighton.artist.common.domain.entity.enums.ApproveStatus;
 import com.prography.lighton.artist.users.infrastructure.repository.ArtistRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
