+++ src/main/java/com/prography/lighton/artist/admin/application/mapper/PendingArtistMapper.java
@@ -2,8 +2,8 @@ package com.prography.lighton.artist.admin.application.mapper;
 
 import com.prography.lighton.artist.admin.presentation.GetArtistApplicationDetailResponseDTO;
 import com.prography.lighton.artist.admin.presentation.GetArtistApplicationListResponseDTO.ArtistApplicationDTO;
-import com.prography.lighton.artist.users.domain.entity.Artist;
-import com.prography.lighton.artist.users.domain.entity.ArtistGenre;
+import com.prography.lighton.artist.common.domain.entity.Artist;
+import com.prography.lighton.artist.common.domain.entity.ArtistGenre;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.genre.infrastructure.cache.GenreCache;
 import java.util.List;
