+++ src/main/java/com/prography/lighton/performance/application/resolver/PerformanceResolver.java
@@ -1,7 +1,7 @@
 package com.prography.lighton.performance.application.resolver;
 
+import com.prography.lighton.artist.common.domain.entity.Artist;
 import com.prography.lighton.artist.users.application.service.ArtistService;
-import com.prography.lighton.artist.users.domain.entity.Artist;
 import com.prography.lighton.genre.application.service.GenreService;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.member.domain.entity.Member;
