+++ src/main/java/com/prography/lighton/performance/application/service/PerformanceService.java
@@ -1,7 +1,7 @@
 package com.prography.lighton.performance.application.service;
 
+import com.prography.lighton.artist.common.domain.entity.Artist;
 import com.prography.lighton.artist.users.application.service.ArtistService;
-import com.prography.lighton.artist.users.domain.entity.Artist;
 import com.prography.lighton.member.domain.entity.Member;
 import com.prography.lighton.performance.application.resolver.PerformanceResolver;
 import com.prography.lighton.performance.domain.entity.Performance;
