+++ src/main/java/com/prography/lighton/performance/domain/entity/Performance.java
@@ -1,11 +1,11 @@
 package com.prography.lighton.performance.domain.entity;
 
-import com.prography.lighton.artist.users.domain.entity.Artist;
-import com.prography.lighton.artist.users.domain.entity.enums.ApproveStatus;
-import com.prography.lighton.artist.users.domain.entity.exception.NotAMasterArtistException;
+import com.prography.lighton.artist.common.domain.entity.Artist;
+import com.prography.lighton.artist.common.domain.entity.exception.NotAMasterArtistException;
 import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.performance.domain.entity.association.PerformanceArtist;
+import com.prography.lighton.performance.domain.entity.enums.ApproveStatus;
 import com.prography.lighton.performance.domain.entity.enums.Seat;
 import com.prography.lighton.performance.domain.entity.enums.Type;
 import com.prography.lighton.performance.domain.entity.vo.Info;
